use events instead of hacky store when new connection established and config loaded
When we make an initial connection or for some reason a new connection
is made that prompts for loading of the inital config, we want to be
able to emit an event out to any frontend components that need to
subscribe to that info. Such as the edit page(s) that will want to set
some default values and update configCopy when a new initial config is
loaded from the ShiftLight.